Removed newmem
