+++ code/act_wiz.c
@@ -59,7 +59,7 @@
 #include "dioextra.h"
 #include "help.h"
 #include "devextra.h"
-#include "newmem.h"
+#include "pstring.h"
 #include "update.h"
 #include "cabal.h"
 #include "comm.h"
@@ -7923,4 +7923,53 @@ void do_racedump(CHAR_DATA *ch, char *argument)
 	}
 
 	fclose(fp);
+}
+
+void do_memtest(CHAR_DATA *ch, char *argument)
+{
+	char buf[MSL];
+	CHAR_DATA *qch;
+	argument = one_argument(argument, buf);
+	RS.Queue.AddToQueue(6, "do_memtest", "do_bash_queue", do_bash_queue, ch, "Calenduil");
+	return;
+
+	//TODO: what the what???
+
+	if(buf[0]=='\0' || !argument || !is_number(argument))
+		return;
+	if(number_percent()> (.2 * get_skill(ch,gsn_rage)))
+	{
+		sprintf(buf,"%f is the value.\n\r",.2 * get_skill(ch,gsn_rage));
+		send_to_char(buf,ch);
+	}
+	return;
+	if(!str_cmp(buf,"dammod"))
+	{
+		for(qch = char_list; qch; qch = qch->next)
+			if(is_npc(qch) && qch->pIndexData->vnum == 3001)
+				qch->dam_mod = atoi(argument);
+		send_to_char("Dammod changed.\n\r",ch);
+	}
+	if(!str_cmp(buf,"hp"))
+	{
+		for(qch = char_list; qch; qch = qch->next)
+			if(is_npc(qch) && qch->pIndexData->vnum == 3001)
+				qch->max_hit = std::max(qch->max_hit - atoi(argument), 100);
+		send_to_char("Maxhit changed.\n\r",ch);
+	}
+	if(!str_cmp(buf,"begin"))
+	{
+		for(qch = char_list; qch; qch = qch->next)
+			if(is_npc(qch) && qch->pIndexData->vnum == 3001 && !is_affected_by(qch, AFF_DETECT_MAGIC))
+				SET_BIT(qch->affected_by, AFF_DETECT_MAGIC);
+		send_to_char("It HAS BEGUN!\n\r",ch);
+	}
+	if(!str_cmp(buf,"end"))
+	{
+		for(qch = char_list; qch; qch = qch->next)
+			if(is_npc(qch) && qch->pIndexData->vnum == 3001 && qch->in_room->vnum > 2399 &&
+			   qch->in_room->vnum < 2801)
+				extract_char(qch, true);
+		send_to_char("It has.. ENDED!\n\r",ch);
+	}
 }
