+++ code/ban.h
@@ -3,7 +3,6 @@
 
 #include "entity/fwd.h"
 #include "entity/limits.h"
-#include "newmem.h"
 
 //	== table layout ==
 //	CREATE TABLE `bans` (
