+++ code/mspec.c
@@ -14,7 +14,7 @@
 #include "act_comm.h"
 #include "act_move.h"
 #include "fight.h"
-#include "newmem.h"
+#include "pstring.h"
 #include "comm.h"
 #include "interp.h"
 #include "db.h"
