+++ code/string.c
@@ -24,7 +24,7 @@
 #include <time.h>
 #include "merc.h"
 #include "string.h"
-#include "newmem.h"
+#include "pstring.h"
 #include "comm.h"
 #include "interp.h"
 #include "db.h"
